agent: extract sandbox secret/token merging helpers
Move redactSandboxSecrets, redactSandboxSecretsResolved, mergeSecretsIntoTokens,
mergeSandboxTokensStructured, mergeSandboxSecrets out of server.ts into
src/server/agent/sandbox-secrets.ts.

Net: server.ts -110 lines (8142 → 8032). No behaviour change.
